Add ETag fields to get/head object
These fields are used to conditionally download or head files in S3 when
you know the MD5 checksum (e.g. if you have a local copy and want to
avoid a redownload). If-None-Match will throw a 304 not modified status
error if the etag matches. Conversely, If-Match will throw a 412
precondition failed if the etag  does not match. Both of these cases are
tested for and documented in the field haddocks.